+++ src/contexts/index.ts
@@ -4,6 +4,7 @@ export { registerOpenaiServiceCommand } from './quickPickSetup/registerOpenaiSer
 export { registerOpenaiSCMCommand } from './scm/registerOpenaiSCMCommand'
 export { registerOpenSettings } from './settings/registerOpenSettings'
 export { registerConversationCommand } from './conversation/registerConversationCommand'
+export { registerEmbeddingCommand } from './embedding/registerEmbeddingCommand'
 
 export { registerConversationsWebviewView } from './views/conversations'
 export { registerEmbeddingView } from './views/embedding'