+++ src/models/index.ts
@@ -1,2 +1,2 @@
 export { default as getSystemPersonas } from './SystemPersonas'
-export { default as QueryResourcePersona } from './QueryResourcePersona'
+export { default as getQueryResourcePersona } from './QueryResourcePersona'