+++ src/services/configurationService/configurationSettingService.ts
@@ -1,61 +1,61 @@
-import { extensions, workspace, version } from 'vscode'
+import { extensions, version } from 'vscode'
 import * as crypto from 'crypto'
 import { SecretStorageService, StatusBarHelper } from '@app/utilities/vscode'
 import {
-  createDebugNotification,
   createErrorNotification,
   createInfoNotification,
   waitFor,
 } from '@app/utilities/node'
-import { IConfigurationService } from '../interfaces/IConfigurationService'
+import ConfigurationService from './configurationService'
+import { IConfigurationSetting } from '@app/interfaces'
 
-export default class ConfigurationService implements IConfigurationService {
-  private static _instance: ConfigurationService
-
-  static init(): void {
-    try {
-      this._instance = new ConfigurationService()
-    } catch (error) {
-      createErrorNotification(error)
+export default class ConfigurationSettingService
+  extends ConfigurationService
+  implements IConfigurationSetting
+{
+  private static _instance: ConfigurationSettingService
+  static get instance(): ConfigurationSettingService {
+    if (!this._instance) {
+      try {
+        this._instance = new ConfigurationSettingService()
+      } catch (error) {
+        createErrorNotification(error)
+      }
     }
-  }
-
-  static get instance(): ConfigurationService {
     return this._instance
   }
 
-  static async loadConfigurationService(config: IConfigurationService) {
+  static async loadConfigurationService({
+    serviceProvider,
+    baseUrl,
+    defaultModel,
+    embeddingModel,
+    azureDeployment,
+    embeddingsDeployment,
+    azureApiVersion,
+  }: {
+    serviceProvider: string
+    baseUrl: string
+    defaultModel: string
+    embeddingModel: string
+    azureDeployment: string
+    embeddingsDeployment: string
+    azureApiVersion: string
+  }) {
     StatusBarHelper.instance.showStatusBarInformation(
       'vscode-openai',
-      'update-cfg'
+      'update-setting-configuration'
     )
-    this._instance.serviceProvider = config.serviceProvider
-    this._instance.baseUrl = config.baseUrl
-    this._instance.defaultModel = config.defaultModel
-    this._instance.embeddingModel = config.embeddingModel
-    this._instance.azureDeployment = config.azureDeployment
-    this._instance.embeddingsDeployment = config.embeddingsDeployment
-    this._instance.azureApiVersion = config.azureApiVersion
+    this.instance.serviceProvider = serviceProvider
+    this.instance.baseUrl = baseUrl
+    this.instance.defaultModel = defaultModel
+    this.instance.embeddingModel = embeddingModel
+    this.instance.azureDeployment = azureDeployment
+    this.instance.embeddingsDeployment = embeddingsDeployment
+    this.instance.azureApiVersion = azureApiVersion
     //Force wait as we need the config to be written
     await waitFor(500, () => false)
     StatusBarHelper.instance.showStatusBarInformation('vscode-openai', '')
-  }
-
-  private getConfigValue<T>(configName: string): T {
-    const ws = workspace.getConfiguration('vscode-openai')
-    return ws.get(configName) as T
-  }
-
-  private setConfigValue<T>(configName: string, value: T): void {
-    const ws = workspace.getConfiguration('vscode-openai')
-    const setAsGlobal = ws.inspect(configName)?.workspaceValue == undefined
-    ws.update(configName, value, setAsGlobal)
-      .then(() => {
-        createDebugNotification(`setting ${configName} ${value}`)
-      })
-      .then(undefined, (err) => {
-        createErrorNotification(`${err}`)
-      })
   }
 
   public get serviceProvider(): string {
@@ -70,6 +70,20 @@ export default class ConfigurationService implements IConfigurationService {
   }
   public set baseUrl(value: string) {
     this.setConfigValue<string>('baseUrl', value)
+  }
+
+  public get defaultModel(): string {
+    return this.getConfigValue<string>('defaultModel')
+  }
+  public set defaultModel(value: string) {
+    this.setConfigValue<string>('defaultModel', value)
+  }
+
+  public get embeddingModel(): string {
+    return this.getConfigValue<string>('embeddingModel')
+  }
+  public set embeddingModel(value: string) {
+    this.setConfigValue<string>('embeddingModel', value)
   }
 
   public get azureDeployment(): string {
@@ -91,24 +105,6 @@ export default class ConfigurationService implements IConfigurationService {
   }
   public set azureApiVersion(value: string) {
     this.setConfigValue<string>('azureApiVersion', value)
-  }
-
-  public get defaultModel(): string {
-    return this.getConfigValue<string>('defaultModel')
-  }
-  public set defaultModel(value: string) {
-    this.setConfigValue<string>('defaultModel', value)
-  }
-
-  public get embeddingModel(): string {
-    return this.getConfigValue<string>('embeddingModel')
-  }
-  public set embeddingModel(value: string) {
-    this.setConfigValue<string>('embeddingModel', value)
-  }
-
-  public get conversationHistory(): number {
-    return this.getConfigValue<number>('conversationHistory')
   }
 
   // host is used for vscode status bar display only
@@ -155,31 +151,35 @@ export default class ConfigurationService implements IConfigurationService {
   }
 
   public get extensionVersion(): string {
-    const extension = extensions.getExtension(
-      'AndrewButson.vscode-openai'
-    )?.packageJSON
-    return extension.version ? extension.version.toString() : 'beta'
+    try {
+      const extension = extensions.getExtension(
+        'AndrewButson.vscode-openai'
+      )?.packageJSON
+      return extension.version ? extension.version.toString() : 'beta'
+    } catch (error) {
+      createErrorNotification(error)
+    }
+    return ''
   }
 
   public static LogConfigurationService(): void {
     try {
-      const instance = this._instance
       const cfgMap = new Map<string, string>()
-
       cfgMap.set('vscode_version', version)
-      cfgMap.set('extension_version', instance.extensionVersion)
-      cfgMap.set('service_provider', instance.serviceProvider)
-      cfgMap.set('host', instance.host)
-      cfgMap.set('base_url', instance.baseUrl)
-      cfgMap.set('inference_model', instance.defaultModel)
-      cfgMap.set('inference_deploy', instance.azureDeployment)
-      cfgMap.set('embeddings_model', instance.embeddingModel)
-      cfgMap.set('embeddings_deploy', instance.embeddingsDeployment)
-      cfgMap.set('az_api_version', instance.azureApiVersion)
-      const convHist = instance.conversationHistory.toString()
-      cfgMap.set('conversation_history', convHist)
+      cfgMap.set('extension_version', this.instance.extensionVersion)
+      cfgMap.set('service_provider', this.instance.serviceProvider)
+      cfgMap.set('host', this.instance.host)
+      cfgMap.set('base_url', this.instance.baseUrl)
+      cfgMap.set('inference_model', this.instance.defaultModel)
+      cfgMap.set('inference_deploy', this.instance.azureDeployment)
+      cfgMap.set('embeddings_model', this.instance.embeddingModel)
+      cfgMap.set('embeddings_deploy', this.instance.embeddingsDeployment)
+      cfgMap.set('az_api_version', this.instance.azureApiVersion)
 
-      createInfoNotification(Object.fromEntries(cfgMap), 'configuration')
+      createInfoNotification(
+        Object.fromEntries(cfgMap),
+        'setting_configuration'
+      )
     } catch (error) {
       createErrorNotification(error)
     }