+++ src/services/index.ts
@@ -1,4 +1,7 @@
-export { default as ConfigurationService } from './configurationService'
+export { default as ConfigurationSettingService } from './configurationService/configurationSettingService'
+export { default as ConfigurationConversationService } from './configurationService/configurationConversationService'
+export { default as ConfigurationEmbeddingService } from './configurationService/configurationEmbeddingService'
+
 export { default as ConversationStorageService } from './conversationStorageService'
 export { default as EmbeddingStorageService } from './embeddingStorageService'
 export { enableServiceFeature, featureVerifyApiKey } from './featureFlagService'