+++ src/utilities/embedding/chunk/searchFileChunks.ts
@@ -2,10 +2,7 @@ import { StatusBarHelper } from '@app/utilities/vscode'
 import { IEmbeddingFileChunk, IEmbeddingFileLite } from '@app/interfaces'
 import { createDebugNotification } from '@app/utilities/node'
 import { createEmbedding } from '@app/utilities/openai'
-
-// This is the minimum cosine similarity score that a file must have with the search query to be considered relevant
-// This is an arbitrary value, and you should vary/ remove this depending on the diversity of your dataset
-const COSINE_SIM_THRESHOLD = 0.72
+import { ConfigurationEmbeddingService } from '@app/services'
 
 // This function takes a search query and a list of files, and returns the chunks of text that are most semantically similar to the query
 export async function searchFileChunks({
@@ -18,6 +15,11 @@ export async function searchFileChunks({
   maxResults: number
 }): Promise<IEmbeddingFileChunk[]> {
   createDebugNotification(`embedding-controller: searchQuery: ${searchQuery}`)
+
+  // This is the minimum cosine similarity score that a file must have with the search query to be considered relevant
+  // This is an arbitrary value, and you should vary/ remove this depending on the diversity of your dataset
+  const COSINE_SIM_THRESHOLD =
+    ConfigurationEmbeddingService.instance.cosineSimilarityThreshold
 
   const searchQueryEmbeddingResponse = await createEmbedding({
     input: searchQuery,
@@ -40,14 +42,22 @@ export async function searchFileChunks({
     // Map each file to an array of chunks with the file name and score
     .flatMap((file) =>
       file.chunks
-        ? file.chunks.map((chunk) => {
+        ? file.chunks.map((chunk, index, array) => {
             // Calculate the dot product between the chunk embedding and the search query embedding
             const dotProduct = chunk.embedding.reduce(
               (sum, val, i) => sum + val * searchQueryEmbedding[i],
               0
             )
+            // Set context by padding text with previous and next segment so that
+            // any hit has more context to answer the question correctly
+            const previousChunkText =
+              index - 1 in array ? array[index - 1]?.text : ''
+            const nextChunkText =
+              index + 1 in array ? array[index + 1]?.text : ''
+            const cloneChunk = { ...chunk }
+            cloneChunk.text = `${previousChunkText}\n${chunk.text}\n${nextChunkText}`
             // Assign the dot product as the score for the chunk
-            return { ...chunk, filename: file.name, score: dotProduct }
+            return { ...cloneChunk, filename: file.name, score: dotProduct }
           })
         : []
     )