+++ src/utilities/openai/api/chatCompletionMessages/chatCompletionRequestMessageEmbedding.ts
@@ -5,13 +5,18 @@ import {
 import { EmbeddingStorageService } from '@app/services'
 import { IConversation, IEmbeddingFileLite } from '@app/interfaces'
 import { searchFileChunks } from '@app/utilities/embedding'
-
-const MAX_FILES_LENGTH = 2000 * 3
-const MAX_RESULTS = 15
+import { StatusBarHelper } from '@app/utilities/vscode'
 
 export async function ChatCompletionRequestMessageEmbedding(
   conversation: IConversation
 ): Promise<ChatCompletionRequestMessage[]> {
+  const MAX_RESULTS = 10
+
+  StatusBarHelper.instance.showStatusBarInformation(
+    'sync~spin',
+    `- search file chunks (${MAX_RESULTS})`
+  )
+
   const chatCompletion: ChatCompletionRequestMessage[] = []
 
   chatCompletion.push({
@@ -22,10 +27,10 @@ export async function ChatCompletionRequestMessageEmbedding(
   const searchQuery =
     conversation.chatMessages[conversation.chatMessages.length - 1].content
 
-  let embeddingFileLites: IEmbeddingFileLite[] = []
+  const embeddingFileLites: Array<IEmbeddingFileLite> = []
   conversation.embeddingId!.forEach((embeddingId) => {
     const embeddingFileLite = EmbeddingStorageService.instance.get(embeddingId)
-    if (embeddingFileLite) embeddingFileLites = [...[embeddingFileLite]]
+    if (embeddingFileLite) embeddingFileLites.push(embeddingFileLite)
   })
 
   const searchFiles = await searchFileChunks({
@@ -34,10 +39,15 @@ export async function ChatCompletionRequestMessageEmbedding(
     maxResults: MAX_RESULTS,
   })
 
+  StatusBarHelper.instance.showStatusBarInformation(
+    'sync~spin',
+    `- found file chunks (${searchFiles.length})`
+  )
+
   const filesString = searchFiles
     .map((searchFiles) => `###\n"${searchFiles.filename}"\n${searchFiles.text}`)
     .join('\n')
-    .slice(0, MAX_FILES_LENGTH)
+    .slice(0)
 
   const content =
     `Question: ${searchQuery}\n\n` + `Files:\n${filesString}\n\n` + `Answer:`
@@ -46,6 +56,8 @@ export async function ChatCompletionRequestMessageEmbedding(
     role: ChatCompletionRequestMessageRoleEnum.Assistant,
     content: content,
   })
+
+  StatusBarHelper.instance.showStatusBarInformation('vscode-openai', '')
 
   return chatCompletion
 }