+++ src/utilities/openai/api/createChatCompletion.ts
@@ -1,7 +1,10 @@
 import { Configuration, OpenAIApi } from 'openai'
 import { BackoffOptions, backOff } from 'exponential-backoff'
 import { StatusBarHelper } from '@app/utilities/vscode'
-import { ConfigurationService } from '@app/services'
+import {
+  ConfigurationConversationService,
+  ConfigurationSettingService,
+} from '@app/services'
 import { IConversation, IMessage } from '@app/interfaces'
 import { errorHandler } from './errorHandler'
 import {
@@ -18,14 +21,14 @@ export async function createChatCompletion(
   try {
     StatusBarHelper.instance.showStatusBarInformation(
       'sync~spin',
-      '- completion'
+      '- build-conversation'
     )
-    const apiKey = await ConfigurationService.instance.getApiKey()
+    const apiKey = await ConfigurationSettingService.instance.getApiKey()
     if (!apiKey) return undefined
 
     const configuration = new Configuration({
       apiKey: apiKey,
-      basePath: ConfigurationService.instance.inferenceUrl,
+      basePath: ConfigurationSettingService.instance.inferenceUrl,
     })
     const openai = new OpenAIApi(configuration)
 
@@ -33,10 +36,16 @@ export async function createChatCompletion(
       ? await ChatCompletionRequestMessageEmbedding(conversation)
       : await ChatCompletionRequestMessageStandard(conversation, responseFormat)
 
-    const requestConfig = await ConfigurationService.instance.getRequestConfig()
+    const requestConfig =
+      await ConfigurationSettingService.instance.getRequestConfig()
+
+    StatusBarHelper.instance.showStatusBarInformation(
+      'sync~spin',
+      '- completion'
+    )
 
     const backoffOptions: BackoffOptions = {
-      numOfAttempts: 20,
+      numOfAttempts: ConfigurationConversationService.instance.numOfAttempts,
       retry: async (_e: any, _attemptNumber: number) => {
         const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms))
         await sleep(1000)
@@ -48,11 +57,13 @@ export async function createChatCompletion(
       () =>
         openai.createChatCompletion(
           {
-            model: ConfigurationService.instance.defaultModel,
+            model: ConfigurationSettingService.instance.defaultModel,
             messages: chatCompletionMessages,
-            temperature: 0.2,
-            frequency_penalty: 0.5,
-            presence_penalty: 0.5,
+            temperature: ConfigurationConversationService.instance.temperature,
+            frequency_penalty:
+              ConfigurationConversationService.instance.frequencyPenalty,
+            presence_penalty:
+              ConfigurationConversationService.instance.presencePenalty,
           },
           requestConfig
         ),